requirements.txt
atomicwrites==1.4.0
attrs==19.1.0
backports.functools-lru-cache==1.6.1
BTrees==4.5.1
CacheControl==0.12.6
cachetools==3.1.1
certifi==2019.3.9
cffi
chardet==3.0.4
Click==7.0
configparser==4.0.2
contextlib2==0.6.0.post1
crcmod==1.7
cycler==0.10.0
enum34==1.1.10
firebase-admin==3.2.1
flasgger==0.9.2
Flask==1.1.1
Flask-Cors==3.0.8
funcsigs==1.0.2
gevent==1.4.0
google-api-core==1.31.3
google-api-python-client==1.12.8
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.2
google-cloud-firestore==1.9.0
google-cloud-storage==1.42.3
google-resumable-media==1.3.3
googleapis-common-protos==1.52.0
greenlet==0.4.15
grpcio==1.41.1
gunicorn==19.9.0
httplib2==0.20.2
idna==2.8
importlib-metadata==2.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.0.1
kiwisolver==1.1.0
locustio==0.11.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==2.2.4
mistune==0.8.4
more-itertools==5.0.0
msgpack==0.6.1
numpy==1.21.4
packaging==20.9
pandas
pathlib==1.0.1
pathlib2==2.3.6
persistent==4.4.3
pg8000==1.12.5
Pillow==6.2.2
pkg-resources==0.0.0
pluggy==0.13.1
praat-parselmouth
protobuf==3.17.3
psycopg2==2.8.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyAudioAnalysis==0.2.5
pycparser==2.19
pydub==0.23.1
pympi-ling==1.69
pyparsing==2.4.6
pyrsistent==0.14.11
pytest==4.6.11
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1
pyzmq==18.0.1
requests==2.21.0
rsa==4.5
scandir==1.10.0
scikit-learn==0.20.3
scipy
seaborn==0.9.0
six==1.14.0
subprocess32==3.5.4
testing.common.database==2.0.3
testing.postgresql==1.3.0
transaction==2.4.0
uritemplate==3.0.1
urllib3==1.24.3
waitress==1.1.0
wcwidth==0.2.5
Werkzeug==0.16.0
zc.lockfile==1.3.0
ZConfig==3.3.0
zipp==1.2.0
ZODB==5.5.1
zodbpickle==1.0.2
zope.interface==4.6.0
